requirements.txt
apimatic-core==0.2.12
apimatic-core-interfaces==0.1.5
apimatic-requests-client-adapter==0.1.6
argcomplete==3.1.6
arrow==1.3.0
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
backcall==0.2.0
beautifulsoup4==4.12.2
build==1.0.3
CacheControl==0.12.14
certifi==2023.11.17
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
colorlog==6.7.0
constantly==15.1.0
crashtest==0.4.1
cryptography==40.0.2
daphne==4.0.0
decorator==5.1.1
deprecation==2.1.0
deptry==0.16.1
distlib==0.3.7
dj-database-url==2.1.0
dj-stripe==2.8.3
Django==5.0.2
django-address==0.2.8
django-appconf==1.0.6
django-autoslug==1.9.9
django-autoslug-field==0.2.3
django-compressor==4.4
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-formtools==2.4.1
django-ipware==6.0.4
django-libsass==0.9
django-livereload==1.7
django-pdb==0.6.2
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
docker==6.1.2
drf-nested-routers==0.93.4
dulwich==0.21.7
enum34==1.1.10
executing==1.2.0
fastjsonschema==2.19.0
filelock==3.13.1
gunicorn==22.0.0
h11==0.14.0
hyperlink==21.0.0
idna==3.6
importlib-metadata==7.0.0
incremental==22.10.0
iniconfig==2.0.0
installer==0.7.0
ipython==8.12.0
jaraco.classes==3.3.0
jedi==0.18.2
Jinja2==3.1.2
jinja2-time==0.2.0
jsonpickle==3.0.2
jsonpointer==2.4
keyring==24.3.0
libsass==0.22.0
make==0.1.6.post2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
more-itertools==10.1.0
msgpack==1.0.7
mysqlclient==2.1.1
outcome==1.3.0.post0
packaging==23.2
parso==0.8.3
pexpect==4.9.0
phonenumbers==8.13.25
phonenumberslite==8.13.26
pickleshare==0.7.5
Pillow==10.1.0
pkginfo==1.9.6
platformdirs==3.11.0
pluggy==1.0.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.15.1
pyOpenSSL==23.1.1
pyproject_hooks==1.0.0
PySocks==1.7.1
pytest==7.3.1
pytest-django==4.5.2
python-dateutil==2.8.2
python-ipware==2.0.2
pytz==2023.3.post1

rapidfuzz==3.5.2
rcssmin==1.1.1
redis==4.5.5
requests==2.31.0
requests-toolbelt==1.0.0
rjsmin==1.2.1
selenium==4.16.0
service-identity==21.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
squareup==33.1.0.20231213
stack-data==0.6.2
stripe==4.2.0
tomlkit==0.12.3
traitlets==5.9.0
trio==0.23.2
trio-websocket==0.11.1
trove-classifiers==2023.11.29

txaio==23.1.1
types-python-dateutil==2.8.19.14
typing_extensions==4.6.2
tzdata==2023.3
urllib3==2.1.0
urwid==2.3.4
userpath==1.9.1
virtualenv==20.25.0
wcwidth==0.2.6
websocket-client==1.5.2
Werkzeug==3.0.1
whitenoise==6.6.0
wsproto==1.2.0
zipp==3.17.0
zope.interface==6.0
